+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * (IRISHub) [\#2918](https://github.com/irisnet/irishub/pull/2918) Adjusting the parameters of the ica module.
 * (IRISHub) [\#2919](https://github.com/irisnet/irishub/pull/2919) Adjusting gov config.
 * (IRISHub) [\#2924](https://github.com/irisnet/irishub/pull/2924) Update evm `AllowUnprotectedTxs` params.
+* (IRISHub) [\#2928](https://github.com/irisnet/irishub/pull/2928) Initialize gov `MinDepositRatio` params.
 * (IRISMod) [\#385](https://github.com/irisnet/irismod/pull/385) Bump cosmos-sdk to v0.47.9-ics-lsm.
 
 ## 2.1.0